DOCS/hal_client/requirements.txt
sounddevice==0.4.6
numpy==1.26.4
webrtcvad==2.0.10
websockets==12.0
simpleaudio==1.0.4
PyYAML==6.0.2
python-dotenv==1.0.1
openwakeword==0.6.0
